Fix comment in keywords.txt
The first line of the file was intended to be a comment but the first few characters were somehow lost. This caused that line to be interpreted as a malformed keyword definition.

The library name in the comment was incorrectly stated as "Audio" rather than the true name of this library: "Brief".